backend/requirements.txt
fastapi
uvicorn[standard]
tensorflow==2.15.0
numpy==1.26.4
Pillow
wandb
